+++ Chord Provider/PDFBuild/Core/PDFBuild+sfSymbol.swift
@@ -19,6 +19,12 @@ extension PDFBuild {
         return PDFBuild.sfSymbol(sfSymbol: sfSymbol.rawValue, fontSize: fontSize, nsColors: [nsColor])
     }
 
+    /// Add an SF Symbol as `NSTextAttachment`
+    /// - Parameters:
+    ///   - sfSymbol: The ``SFSymbol`` to use
+    ///   - fontSize: The size of the font
+    ///   - nsColors: The colors of the icon
+    /// - Returns: An `NSTextAttachment`
     static func sfSymbol(sfSymbol: String, fontSize: Double, nsColors: [NSColor]) -> NSTextAttachment {
         /// Get a large SF symbol and scale it back to expected size or else it will be super blurry
         var config = NSImage.SymbolConfiguration(pointSize: fontSize * 3, weight: .medium, scale: .medium)