+++ Chord Provider/Views/Pickers/FontPicker.swift
@@ -15,7 +15,7 @@ struct FontPicker: View {
     @Binding var options: ConfigOptions.FontOptions
     /// The available fonts of a font-family
     @State private var fontStyles: [FontItem] = []
-
+    /// The currently selected family name
     @State private var familyName: String = ""
     /// The body of the `View`
     var body: some View {